+++ matflow_defdap/snippets/load_EBSD_map.py
@@ -1,27 +1,26 @@
 from pathlib import Path
 
+import numpy as np
 import defdap.ebsd as ebsd
 
 from matflow_defdap import main_func
 
 
 @main_func
-def load_EBSD_map(root_path, ebsd_filename, ebsd_flip_vert,
-                  ebsd_boundary_tol, ebsd_min_grain_size):
+def load_EBSD_map(root_path, ebsd_filename, ebsd_boundary_tol,
+                  ebsd_min_grain_size):
     'Load EBSD map and detect grains.'
 
-    EbsdMap = ebsd.Map(Path(root_path).joinpath(ebsd_filename))
+    ebsd_map = ebsd.Map(Path(root_path).joinpath(ebsd_filename))
 
-    # Flip EBSD map in vertical direction if needed
-    if ebsd_flip_vert:
-        EbsdMap.eulerAngleArray = EbsdMap.eulerAngleArray[:, ::-1, ::-1]
-        EbsdMap.bandContrastArray = EbsdMap.bandContrastArray[::-1, ::-1]
-        EbsdMap.phaseArray = EbsdMap.phaseArray[::-1, ::-1]
+    # check for non-indexed points
+    if np.count_nonzero(ebsd_map.phaseArray == 0) != 0:
+        raise ValueError('EBSD map contains non-indexed points.')
 
-    EbsdMap.buildQuatArray()
+    ebsd_map.buildQuatArray()
 
-    EbsdMap.findBoundaries(boundDef=ebsd_boundary_tol)
-    EbsdMap.findGrains(minGrainSize=ebsd_min_grain_size)
-    EbsdMap.calcGrainAvOris()
+    ebsd_map.findBoundaries(boundDef=ebsd_boundary_tol)
+    ebsd_map.findGrains(minGrainSize=ebsd_min_grain_size)
+    ebsd_map.calcGrainAvOris()
 
-    return EbsdMap
+    return ebsd_map